feat: website-file-breakdown-for-current-selection-002 - Extract workstation selection package
