+++ ui/biome.jsonc
@@ -85,14 +85,14 @@
         "src/features/current-selection/components/provider-session-detail-panel.test.tsx",
         "src/features/current-selection/components/state-node-detail.test.tsx",
         "src/features/current-selection/components/work-item-card.test.tsx",
-        "src/features/current-selection/components/workstation-detail-card.editable-configuration.test.tsx",
-        "src/features/current-selection/components/workstation-detail-card.test.tsx",
-        "src/features/current-selection/components/workstation-prompt-monaco.test.ts",
-        "src/features/current-selection/editable-workstation-prompt-state.test.ts",
-        "src/features/current-selection/hooks/use-editable-workstation-configuration-state.test.tsx",
+        "src/features/current-selection/workstation-selection/components/workstation-detail-card.editable-configuration.test.tsx",
+        "src/features/current-selection/workstation-selection/components/workstation-detail-card.test.tsx",
+        "src/features/current-selection/workstation-selection/components/workstation-prompt-monaco.test.ts",
+        "src/features/current-selection/workstation-selection/editing/editable-workstation-prompt-state.test.ts",
+        "src/features/current-selection/workstation-selection/hooks/use-editable-workstation-configuration-state.test.tsx",
         "src/features/current-selection/hooks/useCurrentSelection.actions.test.ts",
         "src/features/current-selection/hooks/useCurrentSelection.test.tsx",
-        "src/features/current-selection/messages/workstation-detail.test.ts",
+        "src/features/current-selection/workstation-selection/messages/workstation-detail.test.ts",
         "src/features/current-selection/selected-work-dispatch-history-helpers.test.ts",
         "src/features/current-selection/state/executionDetails.test.ts",
         "src/features/current-selection/useCurrentSelection.request-helpers.test.ts",
