+++ ui/src/features/current-selection/components/current-selection-cards.tsx
@@ -2,5 +2,5 @@ export { NoSelectionDetailCard } from "./no-selection-detail-card";
 export { StateNodeDetailCard } from "./state-node-detail";
 export { TerminalWorkSummaryCard } from "./terminal-work-summary-detail";
 export { WorkItemDetailCard } from "./work-item-card";
-export { WorkstationDetailCard } from "./workstation-detail-card";
+export { WorkstationDetailCard } from "../workstation-selection/public";
 export { WorkstationRequestDetailCard } from "./workstation-request-detail";
