+++ ui/src/features/current-selection/components/current-selection-widget.localization.test.tsx
@@ -8,7 +8,7 @@ import { CurrentSelectionWidget } from "./current-selection-widget";
 import { selectWorkItemExecutionDetails } from "../state/executionDetails";
 import { resetSelectionHistoryStore } from "../state/selectionHistoryStore";
 import type { DashboardSelection } from "../state/selection-types";
-import { useSaveEditableWorkstationConfiguration } from "../hooks/use-save-editable-workstation-configuration";
+import { useSaveEditableWorkstationConfiguration } from "../workstation-selection/hooks/use-save-editable-workstation-configuration";
 import type { CurrentSelectionState } from "../hooks/useCurrentSelection";
 import { useCurrentWorkstationPromptTemplateValidation } from "../hooks/useCurrentWorkstationPromptTemplateValidation";
 
@@ -21,11 +21,11 @@ vi.mock("../../current-factory-definition/public", async () => {
   };
 });
 
-vi.mock("../hooks/use-save-editable-workstation-configuration", () => ({
+vi.mock("../workstation-selection/hooks/use-save-editable-workstation-configuration", () => ({
   useSaveEditableWorkstationConfiguration: vi.fn(),
 }));
 
-vi.mock("../hooks/useCurrentWorkstationPromptTemplateValidation", () => ({
+vi.mock("../workstation-selection/hooks/useCurrentWorkstationPromptTemplateValidation", () => ({
   useCurrentWorkstationPromptTemplateValidation: vi.fn(),
 }));
