+++ ui/src/features/current-selection/components/detail-card-types.ts
@@ -1,1 +1,2 @@
 export * from "../base/components/detail-card-types";
+export * from "../workstation-selection/lib/detail-card-types";
