+++ ui/src/features/current-selection/workstation-selection/components/workstation-detail-card.lifecycle.test.tsx
@@ -1,7 +1,7 @@
 import { fireEvent, render, screen, within } from "@testing-library/react";
-import type { DashboardWorkstationRequest } from "../../../api/dashboard/types";
-import { semanticWorkflowDashboardSnapshot } from "../../../components/dashboard/test-fixtures";
-import { DETAIL_CARD_NOW } from "./detail-card-test-helpers";
+import type { DashboardWorkstationRequest } from "../../../../api/dashboard/types";
+import { semanticWorkflowDashboardSnapshot } from "../../../../components/dashboard/test-fixtures";
+import { DETAIL_CARD_NOW } from "../../base/components/detail-card-test-helpers";
 import { WorkstationDetailCard } from "./workstation-detail-card";
 
 function requireValue<T>(value: T | null | undefined, message: string): T {
