+++ ui/src/features/current-selection/workstation-selection/components/workstation-detail-card.test.tsx
@@ -5,9 +5,9 @@ import {
   DASHBOARD_SECTION_HEADING_CLASS,
   DASHBOARD_SUPPORTING_CODE_CLASS,
   DASHBOARD_SUPPORTING_TEXT_CLASS,
-} from "../../../components/ui/dashboard-typography";
-import { semanticWorkflowDashboardSnapshot } from "../../../components/dashboard/test-fixtures";
-import { DETAIL_CARD_NOW } from "./detail-card-test-helpers";
+} from "../../../../components/ui/dashboard-typography";
+import { semanticWorkflowDashboardSnapshot } from "../../../../components/dashboard/test-fixtures";
+import { DETAIL_CARD_NOW } from "../../base/components/detail-card-test-helpers";
 import { WorkstationDetailCard } from "./workstation-detail-card";
 
 function requireValue<T>(value: T | null | undefined, message: string): T {
