+++ ui/src/features/current-selection/workstation-selection/components/workstation-prompt-editor.test.tsx
