+++ ui/src/features/current-selection/workstation-selection/components/workstation-prompt-field.tsx
@@ -2,16 +2,16 @@ import {
   DASHBOARD_BODY_TEXT_CLASS,
   DASHBOARD_SUPPORTING_LABEL_CLASS,
   DASHBOARD_SUPPORTING_TEXT_CLASS,
-} from "../../../components/ui/dashboard-typography";
-import { cn } from "../../../lib/cn";
+} from "../../../../components/ui/dashboard-typography";
+import { cn } from "../../../../lib/cn";
 import type { getWorkstationDetailMessages } from "../messages/workstation-detail";
 import {
   CURRENT_SELECTION_ALERT_PANEL_CLASS,
   CURRENT_SELECTION_CODE_SUBTLE_CLASS,
   CURRENT_SELECTION_FIELD_PANEL_CLASS,
   CURRENT_SELECTION_NOTICE_SUBTLE_CLASS,
-} from "./detail-card-shared";
-import type { WorkstationDetailCardProps } from "./detail-card-types";
+} from "../../base/components/detail-card-shared";
+import type { WorkstationDetailCardProps } from "../lib/detail-card-types";
 import { WorkstationPromptEditor } from "./workstation-prompt-editor";
 
 export function EditableConfigurationPromptInput({
