+++ ui/src/features/current-selection/workstation-selection/components/workstation-prompt-monaco-language.ts
