+++ ui/src/features/current-selection/workstation-selection/components/workstation-prompt-monaco.test.ts
