+++ ui/src/features/current-selection/workstation-selection/editing/editable-workstation-overwrite-fields.ts
@@ -1,6 +1,6 @@
-import { formatList } from "../../../components/ui/formatters";
-import type { EditableWorkstationDraft } from "../../current-factory-definition/lib/workstation-editable-values";
-import type { EditableWorkstationOverwriteField } from "../components/detail-card-types";
+import { formatList } from "../../../../components/ui/formatters";
+import type { EditableWorkstationDraft } from "../../../current-factory-definition/lib/workstation-editable-values";
+import type { EditableWorkstationOverwriteField } from "../lib/detail-card-types";
 import type { WorkstationDetailMessages } from "../messages/workstation-detail-types";
 
 export function resolveEditableWorkstationOverwriteFields(
