+++ ui/src/features/current-selection/workstation-selection/editing/editable-workstation-prompt-state.test.ts
