+++ ui/src/features/current-selection/workstation-selection/hooks/use-editable-workstation-configuration-state.test.tsx
@@ -1,9 +1,9 @@
 import { act, renderHook, waitFor } from "@testing-library/react";
 
-import { semanticWorkflowDashboardSnapshot } from "../../../components/dashboard/test-fixtures";
-import type { CurrentFactoryDocument } from "../../../api/current-factory-definition";
-import { useCurrentFactoryDocument } from "../../current-factory-definition/public";
-import type { DashboardSelection } from "../state/selection-types";
+import { semanticWorkflowDashboardSnapshot } from "../../../../components/dashboard/test-fixtures";
+import type { CurrentFactoryDocument } from "../../../../api/current-factory-definition";
+import { useCurrentFactoryDocument } from "../../../current-factory-definition/public";
+import type { DashboardSelection } from "../../base/state/selection-types";
 import {
   useEditableWorkstationConfigurationState,
   validateEditableWorkstationDraft,
@@ -11,8 +11,8 @@ import {
 import { useCurrentWorkstationPromptTemplateContract } from "./useCurrentWorkstationPromptTemplateContract";
 import { useCurrentWorkstationPromptTemplateValidation } from "./useCurrentWorkstationPromptTemplateValidation";
 
-vi.mock("../../current-factory-definition/public", async () => {
-  const actual = await vi.importActual("../../current-factory-definition/public");
+vi.mock("../../../current-factory-definition/public", async () => {
+  const actual = await vi.importActual("../../../current-factory-definition/public");
 
   return {
     ...actual,
