+++ ui/src/features/current-selection/workstation-selection/hooks/use-editable-workstation-configuration-state.ts
@@ -1,0 +1,472 @@
+import { useEffect, useMemo, useState } from "react";
+
+import type { DashboardWorkstationNode } from "../../../../api/dashboard/types";
+import { useCurrentFactoryDocument } from "../../../current-factory-definition/public";
+import {
+  applyEditableWorkstationDraft,
+  type EditableWorkstationDraft,
+  editableWorkstationDraftFromValues,
+  resolveEditableWorkstationValues,
+} from "../../../current-factory-definition/lib/workstation-editable-values";
+import {
+  workstationBehaviorRequiresPrompt,
+  workerSupportsPollerBehavior,
+  type EditableWorkstationBehavior,
+} from "../../../current-factory-definition/lib/workstation-behavior";
+import type {
+  EditableWorkstationConfigurationState,
+  EditableWorkstationPromptHelpState,
+  EditableWorkstationPromptValidationState,
+  EditableWorkstationValidationErrors,
+  EditableWorkstationWorkerOptionsState,
+} from "../lib/detail-card-types";
+import {
+  resolvePromptHelpState,
+  resolvePromptValidationState,
+} from "../editing/editable-workstation-prompt-state";
+import { resolveEditableWorkstationOverwriteFields } from "../editing/editable-workstation-overwrite-fields";
+import {
+  getWorkstationDetailMessages,
+  type WorkstationDetailMessages,
+} from "../messages/workstation-detail";
+import type { RunnerID } from "../editing/runner-metadata";
+import type { DashboardSelection } from "../../base/state/selection-types";
+import { useCurrentWorkstationPromptTemplateContract } from "./useCurrentWorkstationPromptTemplateContract";
+import { useCurrentWorkstationPromptTemplateValidation } from "./useCurrentWorkstationPromptTemplateValidation";
+
+interface EditableWorkstationSessionState {
+  draft: EditableWorkstationDraft;
+  latestDefinitionDraft: EditableWorkstationDraft;
+  selectionKey: string;
+  sessionStartDraft: EditableWorkstationDraft;
+}
+
+export function useEditableWorkstationConfigurationState(
+  selection: DashboardSelection | null,
+  selectedNode: DashboardWorkstationNode | null,
+  locale?: string | null,
+): EditableWorkstationConfigurationState | undefined {
+  const isNodeSelection = selection?.kind === "node" && selectedNode != null;
+  const messages = getWorkstationDetailMessages(locale);
+  const editableDefinition = useCurrentFactoryDocument(isNodeSelection);
+  const { selectedEditableValues, sessionState, setSessionState } =
+    useEditableWorkstationSession(
+      editableDefinition.data,
+      selectedNode,
+      selection,
+    );
+  const promptTemplateContract = useCurrentWorkstationPromptTemplateContract(
+    selectedEditableValues?.workstationName,
+    isNodeSelection && selectedEditableValues != null,
+  );
+  const shouldValidatePrompt =
+    isNodeSelection &&
+    sessionState != null &&
+    workstationBehaviorRequiresPrompt(sessionState.draft.behavior);
+  const promptValidation = useCurrentWorkstationPromptTemplateValidation(
+    selectedEditableValues?.workstationName,
+    sessionState?.draft.prompt,
+    shouldValidatePrompt && selectedEditableValues != null,
+  );
+
+  if (!isNodeSelection) {
+    return undefined;
+  }
+
+  if (editableDefinition.isPending) {
+    return { status: "loading" };
+  }
+
+  if (editableDefinition.isError) {
+    return {
+      errorMessage: editableDefinition.error.message,
+      status: "error",
+    };
+  }
+
+  if (!editableDefinition.data || !selectedEditableValues || !sessionState) {
+    return {
+      message: messages.editableConfigurationEmpty,
+      status: "empty",
+    };
+  }
+
+  const promptHelpState = resolvePromptHelpState(
+    promptTemplateContract,
+    messages,
+  );
+  const promptValidationState = resolvePromptValidationState(
+    promptValidation,
+    sessionState.draft.prompt,
+    messages,
+  );
+
+  const resolvedValidationErrors = validateEditableWorkstationDraft(
+    sessionState.draft,
+    selectedEditableValues,
+    promptValidationState,
+    messages,
+  );
+
+  return buildReadyEditableWorkstationConfigurationState({
+    editableDefinition: editableDefinition.data,
+    messages,
+    promptHelpState,
+    promptValidationState,
+    resolvedValidationErrors,
+    selectedEditableValues,
+    selectedNode,
+    sessionState,
+    setSessionState,
+  });
+}
+
+function useEditableWorkstationSession(
+  editableDefinition: ReturnType<
+    typeof useCurrentFactoryDocument
+  >["data"],
+  selectedNode: DashboardWorkstationNode | null,
+  selection: DashboardSelection | null,
+) {
+  const selectedEditableValues = useMemo(() => {
+    if (selection?.kind !== "node" || !selectedNode || !editableDefinition) {
+      return null;
+    }
+
+    return resolveEditableWorkstationValues(editableDefinition, selectedNode);
+  }, [editableDefinition, selectedNode, selection]);
+  const selectionKey =
+    selection?.kind === "node" && selectedNode
+      ? `${selectedNode.node_id}:${selectedNode.transition_id}:${selectedNode.workstation_name}`
+      : null;
+  const [sessionState, setSessionState] =
+    useState<EditableWorkstationSessionState | null>(null);
+
+  useEffect(() => {
+    setSessionState((currentState) =>
+      syncEditableWorkstationSession(
+        currentState,
+        selectedEditableValues,
+        selectionKey,
+      ),
+    );
+  }, [selectedEditableValues, selectionKey]);
+
+  return {
+    selectedEditableValues,
+    sessionState,
+    setSessionState,
+  };
+}
+
+export function validateEditableWorkstationDraft(
+  draft: EditableWorkstationDraft,
+  selectedEditableValues?: ReturnType<typeof resolveEditableWorkstationValues>,
+  promptValidationState: EditableWorkstationPromptValidationState = {
+    status: "idle",
+  },
+  messages: Pick<
+    WorkstationDetailMessages,
+    | "editableConfigurationPromptRequired"
+    | "editableConfigurationPromptValidationLoading"
+    | "editableConfigurationPromptValidationErrorPrefix"
+    | "editableConfigurationPromptDiagnosticsSummary"
+    | "editableConfigurationBehaviorPollerWorkerUnsupported"
+    | "editableConfigurationWorkerRequired"
+    | "editableConfigurationWorkerUnavailable"
+  > = getWorkstationDetailMessages(undefined),
+): EditableWorkstationValidationErrors {
+  const validationErrors: EditableWorkstationValidationErrors = {};
+  const promptIsRequired = workstationBehaviorRequiresPrompt(draft.behavior);
+
+  if (draft.workerName.trim().length === 0) {
+    validationErrors.workerName = messages.editableConfigurationWorkerRequired;
+  } else if (
+    selectedEditableValues &&
+    !selectedEditableValues.workerOptions.includes(draft.workerName)
+  ) {
+    validationErrors.workerName =
+      messages.editableConfigurationWorkerUnavailable;
+  }
+  if (
+    draft.behavior === "POLLER" &&
+    selectedEditableValues &&
+    !workerSupportsPollerBehavior(
+      draft.workerName.trim().length === 0
+        ? null
+        : {
+            type: selectedEditableValues.workerTypeByName[draft.workerName],
+          },
+    )
+  ) {
+    validationErrors.behavior =
+      messages.editableConfigurationBehaviorPollerWorkerUnsupported;
+  }
+
+  if (promptIsRequired && draft.prompt.trim().length === 0) {
+    validationErrors.prompt = messages.editableConfigurationPromptRequired;
+  } else if (promptIsRequired && promptValidationState.status === "loading") {
+    validationErrors.prompt =
+      messages.editableConfigurationPromptValidationLoading;
+  } else if (promptIsRequired && promptValidationState.status === "error") {
+    validationErrors.prompt = `${messages.editableConfigurationPromptValidationErrorPrefix} ${promptValidationState.errorMessage}`;
+  } else if (
+    promptIsRequired &&
+    promptValidationState.status === "ready" &&
+    promptValidationState.diagnostics.length > 0
+  ) {
+    validationErrors.prompt =
+      messages.editableConfigurationPromptDiagnosticsSummary;
+  }
+
+  return validationErrors;
+}
+
+export function hasEditableWorkstationValidationErrors(
+  validationErrors: EditableWorkstationValidationErrors,
+): boolean {
+  return Boolean(
+    validationErrors.behavior ||
+      validationErrors.prompt ||
+      validationErrors.runnerName ||
+      validationErrors.workerName,
+  );
+}
+
+function areEditableDraftsEqual(
+  left: EditableWorkstationDraft,
+  right: EditableWorkstationDraft,
+): boolean {
+  return (
+    left.behavior === right.behavior &&
+    left.prompt === right.prompt &&
+    left.runnerName === right.runnerName &&
+    left.workerName === right.workerName
+  );
+}
+
+function resolveWorkerOptionsState(
+  draft: EditableWorkstationDraft,
+  selectedEditableValues: ReturnType<typeof resolveEditableWorkstationValues>,
+  messages: Pick<
+    WorkstationDetailMessages,
+    | "editableConfigurationEmpty"
+    | "editableConfigurationWorkerMissing"
+    | "editableConfigurationWorkerOptionsEmpty"
+  >,
+): EditableWorkstationWorkerOptionsState {
+  if (!selectedEditableValues) {
+    return {
+      message: messages.editableConfigurationEmpty,
+      status: "error",
+    };
+  }
+
+  if (selectedEditableValues.workerOptions.length === 0) {
+    return {
+      message: messages.editableConfigurationWorkerOptionsEmpty,
+      status: "empty",
+    };
+  }
+
+  if (!selectedEditableValues.workerOptions.includes(draft.workerName)) {
+    return {
+      message: messages.editableConfigurationWorkerMissing,
+      status: "error",
+    };
+  }
+
+  return {
+    options: selectedEditableValues.workerOptions,
+    status: "ready",
+  };
+}
+
+function buildReadyEditableWorkstationConfigurationState({
+  editableDefinition,
+  messages,
+  promptHelpState,
+  promptValidationState,
+  resolvedValidationErrors,
+  selectedEditableValues,
+  selectedNode,
+  sessionState,
+  setSessionState,
+}: {
+  editableDefinition: NonNullable<
+    ReturnType<typeof useCurrentFactoryDocument>["data"]
+  >;
+  messages: WorkstationDetailMessages;
+  promptHelpState: EditableWorkstationPromptHelpState;
+  promptValidationState: EditableWorkstationPromptValidationState;
+  resolvedValidationErrors: EditableWorkstationValidationErrors;
+  selectedEditableValues: NonNullable<
+    ReturnType<typeof resolveEditableWorkstationValues>
+  >;
+  selectedNode: DashboardWorkstationNode;
+  sessionState: EditableWorkstationSessionState;
+  setSessionState: (
+    updater: (
+      currentState: EditableWorkstationSessionState | null,
+    ) => EditableWorkstationSessionState | null,
+  ) => void;
+}) {
+  const pendingFactoryDefinition = hasEditableWorkstationValidationErrors(
+    resolvedValidationErrors,
+  )
+    ? null
+    : applyEditableWorkstationDraft(
+        editableDefinition,
+        selectedNode,
+        sessionState.draft,
+      );
+
+  return {
+    baseVersion: editableDefinition.version,
+    draft: sessionState.draft,
+    hasValidationErrors: hasEditableWorkstationValidationErrors(
+      resolvedValidationErrors,
+    ),
+    initialValues: selectedEditableValues,
+    isDirty: !areEditableDraftsEqual(
+      sessionState.draft,
+      sessionState.sessionStartDraft,
+    ),
+    markChangesSaved: () => {
+      setSessionState((currentState) =>
+        currentState
+          ? {
+              ...currentState,
+              latestDefinitionDraft: currentState.draft,
+              sessionStartDraft: currentState.draft,
+            }
+          : currentState,
+      );
+    },
+    onPromptChange: (value: string) => {
+      setSessionState((currentState) =>
+        currentState
+          ? {
+              ...currentState,
+              draft: {
+                ...currentState.draft,
+                prompt: value,
+              },
+            }
+          : currentState,
+      );
+    },
+    onResetToLatest: () => {
+      setSessionState((currentState) =>
+        currentState
+          ? {
+              ...currentState,
+              draft: currentState.latestDefinitionDraft,
+              sessionStartDraft: currentState.latestDefinitionDraft,
+            }
+          : currentState,
+      );
+    },
+    onBehaviorChange: (value: EditableWorkstationBehavior) => {
+      setSessionState((currentState) =>
+        currentState
+          ? {
+              ...currentState,
+              draft: {
+                ...currentState.draft,
+                behavior: value,
+              },
+            }
+          : currentState,
+      );
+    },
+    onRunnerChange: (value: RunnerID | null) => {
+      setSessionState((currentState) =>
+        currentState
+          ? {
+              ...currentState,
+              draft: {
+                ...currentState.draft,
+                runnerName: value,
+              },
+            }
+          : currentState,
+      );
+    },
+    onWorkerChange: (value: string) => {
+      setSessionState((currentState) =>
+        currentState
+          ? {
+              ...currentState,
+              draft: {
+                ...currentState.draft,
+                workerName: value,
+              },
+            }
+          : currentState,
+      );
+    },
+    overwriteFieldNames: resolveEditableWorkstationOverwriteFields(
+      sessionState.sessionStartDraft,
+      sessionState.draft,
+      sessionState.latestDefinitionDraft,
+    ),
+    pendingFactoryDefinition,
+    promptDiagnostics:
+      promptValidationState.status === "ready"
+        ? promptValidationState.diagnostics
+        : [],
+    promptHelpState,
+    promptValidationState,
+    status: "ready" as const,
+    validationErrors: resolvedValidationErrors,
+    workerOptionsState: resolveWorkerOptionsState(
+      sessionState.draft,
+      selectedEditableValues,
+      messages,
+    ),
+  };
+}
+
+function syncEditableWorkstationSession(
+  currentState: EditableWorkstationSessionState | null,
+  selectedEditableValues: ReturnType<typeof resolveEditableWorkstationValues>,
+  selectionKey: string | null,
+): EditableWorkstationSessionState | null {
+  if (!selectionKey || !selectedEditableValues) {
+    return null;
+  }
+
+  const initialDraft = editableWorkstationDraftFromValues(
+    selectedEditableValues,
+  );
+  if (!currentState || currentState.selectionKey !== selectionKey) {
+    return {
+      draft: initialDraft,
+      latestDefinitionDraft: initialDraft,
+      selectionKey,
+      sessionStartDraft: initialDraft,
+    };
+  }
+
+  if (
+    areEditableDraftsEqual(currentState.draft, currentState.sessionStartDraft)
+  ) {
+    return {
+      draft: initialDraft,
+      latestDefinitionDraft: initialDraft,
+      selectionKey,
+      sessionStartDraft: initialDraft,
+    };
+  }
+
+  return areEditableDraftsEqual(
+    currentState.latestDefinitionDraft,
+    initialDraft,
+  )
+    ? currentState
+    : {
+        ...currentState,
+        latestDefinitionDraft: initialDraft,
+      };
+}
