+++ ui/src/features/current-selection/workstation-selection/messages/workstation-detail-enums.ts
@@ -1,0 +1,164 @@
+import {
+  localizeEnumLabel,
+  type LocalizedMessageCatalog,
+  resolveLocalizedMessages,
+} from "../../../../i18n";
+
+export interface WorkstationDetailEnumMessages {
+  localizeProviderSessionKind: (value: string) => string;
+  localizeWorkstationBehavior: (value: string) => string;
+  localizeWorkstationKind: (value: string) => string;
+}
+
+const workstationDetailEnumMessagesByLocale = {
+  en: {
+    localizeProviderSessionKind: (value: string) =>
+      localizeEnumLabel({
+        category: "kind",
+        labels: {
+          path: "Path",
+          session_id: "Session ID",
+        },
+        locale: "en",
+        value,
+      }),
+    localizeWorkstationBehavior: (value: string) =>
+      localizeEnumLabel({
+        category: "kind",
+        labels: {
+          CRON: "Cron",
+          POLLER: "Poller",
+          REPEATER: "Repeater",
+          STANDARD: "Standard",
+        },
+        locale: "en",
+        value,
+      }),
+    localizeWorkstationKind: (value: string) =>
+      localizeEnumLabel({
+        category: "kind",
+        labels: {
+          cron: "Cron",
+          poller: "Poller",
+          repeater: "Repeater",
+          standard: "Standard",
+        },
+        locale: "en",
+        value,
+      }),
+  },
+  ja: {
+    localizeProviderSessionKind: (value: string) =>
+      localizeEnumLabel({
+        category: "kind",
+        labels: {
+          path: "パス",
+          session_id: "セッション ID",
+        },
+        locale: "ja",
+        value,
+      }),
+    localizeWorkstationBehavior: (value: string) =>
+      localizeEnumLabel({
+        category: "kind",
+        labels: {
+          CRON: "Cron",
+          POLLER: "ポーラー",
+          REPEATER: "リピーター",
+          STANDARD: "標準",
+        },
+        locale: "ja",
+        value,
+      }),
+    localizeWorkstationKind: (value: string) =>
+      localizeEnumLabel({
+        category: "kind",
+        labels: {
+          cron: "Cron",
+          poller: "ポーラー",
+          repeater: "リピーター",
+          standard: "標準",
+        },
+        locale: "ja",
+        value,
+      }),
+  },
+  ko: {
+    localizeProviderSessionKind: (value: string) =>
+      localizeEnumLabel({
+        category: "kind",
+        labels: {
+          path: "경로",
+          session_id: "세션 ID",
+        },
+        locale: "ko",
+        value,
+      }),
+    localizeWorkstationBehavior: (value: string) =>
+      localizeEnumLabel({
+        category: "kind",
+        labels: {
+          CRON: "Cron",
+          POLLER: "폴러",
+          REPEATER: "반복기",
+          STANDARD: "표준",
+        },
+        locale: "ko",
+        value,
+      }),
+    localizeWorkstationKind: (value: string) =>
+      localizeEnumLabel({
+        category: "kind",
+        labels: {
+          cron: "Cron",
+          poller: "폴러",
+          repeater: "반복기",
+          standard: "표준",
+        },
+        locale: "ko",
+        value,
+      }),
+  },
+  "zh-CN": {
+    localizeProviderSessionKind: (value: string) =>
+      localizeEnumLabel({
+        category: "kind",
+        labels: {
+          path: "路径",
+          session_id: "会话 ID",
+        },
+        locale: "zh-CN",
+        value,
+      }),
+    localizeWorkstationBehavior: (value: string) =>
+      localizeEnumLabel({
+        category: "kind",
+        labels: {
+          CRON: "Cron",
+          POLLER: "轮询器",
+          REPEATER: "重复器",
+          STANDARD: "标准",
+        },
+        locale: "zh-CN",
+        value,
+      }),
+    localizeWorkstationKind: (value: string) =>
+      localizeEnumLabel({
+        category: "kind",
+        labels: {
+          cron: "Cron",
+          poller: "轮询器",
+          repeater: "重复器",
+          standard: "标准",
+        },
+        locale: "zh-CN",
+        value,
+      }),
+  },
+} satisfies LocalizedMessageCatalog<WorkstationDetailEnumMessages>;
+
+export function getWorkstationDetailEnumMessages(
+  locale?: string | null,
+): WorkstationDetailEnumMessages {
+  return resolveLocalizedMessages(workstationDetailEnumMessagesByLocale, locale);
+}
